# add CropAndResize layer for CUDA backend