@YashasSamaga 2019-12-05
Is there a test for CropAndResize layer? I am not able to find one.