@tompollok 2019-12-05
Have you tried MaskRCNN with this patch?

Last time without CUDA CropAndResize you used 7700HQ and GTX 1050 and your results were:

Inception v2 Mask RCNN
OCV CPU Time:   3280ms
CUDA Total Time: 407ms
Relative Error >> Total: 0.787171, Average: 1.94363e-07, Max: 1.65757e-06