@YashasSamaga 2019-12-05
@tompollok 

Results below are from the same device configuration (OS is Ubuntu 18.04 instead of Windows).

```
Inception v2 Mask RCNN
OCV CPU Time:   1090ms
IE CPU Time:    583ms
CUDA Total Time: 120.399ms
Relative Error >> Total: 0.63299, Average: 3.12588e-07, Max: 2.26777e-06


ResNet50 Faster RCNN
OCV CPU Time:   916ms
IE CPU Time:    586ms
CUDA Total Time: 133.605ms
Relative Error >> Total: 0.000495682, Average: 7.08118e-07, Max: 2.37544e-05


Inception v2 Faster RCNN
OCV CPU Time:   378ms
IE CPU Time:    251ms
CUDA Total Time: 75.324ms
Relative Error >> Total: 0.000538239, Average: 7.68913e-07, Max: 1.70529e-05
```