@tompollok 2019-12-06
Didnt you say that using CPU fallback results in copying data back and forth from GPU and CPU?
Also there has been some optimization work lately, i dont know if and how that affects the opencv cpu backend. Id say running the latest commit on your windows machine would allow to see if there is actually a big difference on these platforms. Anyways, great job!

Btw: are there still many cpu fallbacks or are these models listed above now running completely on CUDA backend?

Would be kind of interesting to have some debug function that shows the backend forwarding path for a model such that its easy to see which layers are not on the same (cuda) backend