@YashasSamaga 2019-12-06
@tompollok 

Set `OPENCV_LOG_LEVEL` environment variable to `INFO` an you'll be able to see which layers are using fallbacks.

```
[ INFO:0] global /home/yashas/Desktop/gsoc/opencv/modules/dnn/src/dnn.cpp (2204) initCUDABackend CUDA backend will fallback to the CPU implementation for the layer "_input" of type __NetInputLayer__

[ INFO:0] global /home/yashas/Desktop/gsoc/opencv/modules/dnn/src/dnn.cpp (2204) initCUDABackend CUDA backend will fallback to the CPU implementation for the layer "detection_out" of type DetectionOutput

[ INFO:0] global /home/yashas/Desktop/gsoc/opencv/modules/dnn/src/dnn.cpp (2204) initCUDABackend CUDA backend will fallback to the CPU implementation for the layer "detection_out_final" of type DetectionOutput
```

The only missing layer now is DetectionOutput. By the nature of computations involved, I don't think it's worth porting it to GPU fully. It might help to move part of DetectionOutput to GPU and perform the final steps such as NMS on the CPU (like the way it's done for region layer). I'll have to dig through the code and think about it.

Other optimizations are possible:
1. Fuse ReLU with convolution
2. Skip concat (by directly writing to the concat output block by its parent layers)
3. DetectionOutput involves 3 device to host copy before the layer starts working. Two of the inputs are available much before DetectionOutput. The device-to-host copy can be started as early as possible for those two inputs. So only one D2H transfer would be left just before DetectionOutput.