@dkurt 2019-12-09
> Is there a test for CropAndResize layer? I am not able to find one.
> 
> Otherwise, the PR is ready and not WIP.

@YashasSamaga. thanks for this feature! There is no single layer test but there are tests for Faster R-CNNs with it.