@YashasSamaga 2019-12-09
@dkurt I just realized after the notification about this PR's approval that maybe the optimizations added in PR16097 to the resize kernel is applicable here too. I would like to profile and check. If it doesn't cause trouble and you're ok with it, I can add a commit to this PR to optimize the kernel. Otherwise, I will make another PR with optimization.