@JTzhuang 2020-05-22
@dkurt @YashasSamaga      I encountered a strange problem, When I tested with the following code, I found that the inference time of the model is different, time1 = 700ms, time2 = 3ms. And when I run this code on CPU, the inference time = 300ms, It so weired.

`
double start = getTickCount();
net2.forward(outs, outNames);
double time1 = 1000*(getTickCount() - start) / getTickFrequency();
cout << time1  << endl;

vector<double> layersTimings2;
double freq1 = getTickFrequency() / 1000;
double time2 = net2.getPerfProfile(layersTimings2) / freq1;
cout << time2 << endl;
`