@YashasSamaga 2020-05-22
@JTzhuang The first forward pass has initialization cost. You should ignore it. The subsequent forward passes should be faster.