@JTzhuang 2020-05-22
@YashasSamaga The first forward pass is 800ms, and the subsequent forward are about 700ms.