@JTzhuang 2020-05-22
@YashasSamaga 
https://drive.google.com/open?id=1Wsx4JOKvn6Xn2Rr7m0VX2HTXEvUBvWm3  
This is my code and .pb file, you can test it on CPU and GPU.