@YashasSamaga 2020-05-22
@JTzhuang Sorry, I misread your initial post. The CUDA backend doesn't support `getPerfProfile`. If you need to check layerwise timings, you can use NVIDIA Nsight Systems. It provides nice visualizations and easy-to-use GUI interface.

It might be difficult to identify what operations correspond to which layer. CUDA allows profiling data to be annotated in the code using NVTX. The CUDA backend currently doesn't have any profiler annotation capabilities. If you think this could be a useful feature, please open a feature request issue.

**CPU:**
<details>
<summary>CPU output</summary>

```
0.667441[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
653.698
0.198471[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
190.316
0.197608[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
189.356
0.197534[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
188.999
0.19357[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
184.509
0.195504[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
187.31
0.194946[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
186.271
0.197041[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
188.171
0.196134[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
186.883
0.19249[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
184.133
0.19389[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
184.95
0.195315[1.318402e-05, 1.318402e-05, 0.99996042, 1.318402e-05]
186.562

```
</details>

**CUDA:**

<details>
<summary>CUDA output</summary>
0.493976[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
3.0786
0.522221[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.73693
0.52053[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
3.00782
0.52222[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.77092
0.525574[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.69902
0.523383[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
3.3333
0.521101[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.81944
0.528263[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.75951
0.528537[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.76679
0.529091[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
2.72331
0.54491[1.3184058e-05, 1.3184058e-05, 0.99996042, 1.3184058e-05]
3.07419
</details>