@JTzhuang 2020-05-22
@YashasSamaga So you think the inference cost on GPU is right?  It's obvious that the inference time on GPU is much slower than CPU.      
Is there something wrong with my testing method.