@YashasSamaga 2020-05-22
Without `color_input` set.

```
Test Network
[OCV CPU]
        init >> 185.475ms
        inference >> min = 178.937ms, max = 185.759ms, mean = 182.892ms, stddev = 1.91519ms
[CUDA FP32]
        init >> 491.033ms
        inference >> min = 41.773ms, max = 42.021ms, mean = 41.9014ms, stddev = 0.0698771ms
```

With color input:

```
Test Network
[OCV CPU]
        init >> 192.439ms
        inference >> min = 191.679ms, max = 198.518ms, mean = 194.69ms, stddev = 2.36098ms
[CUDA FP32]
        init >> 486.104ms
        inference >> min = 532.097ms, max = 547.586ms, mean = 538.318ms, stddev = 4.02725ms
```

I tried to view your model in [netron](https://lutzroeder.github.io/netron/). No layer seems to be using `color_input`. I need to find to out why giving  `color_input` kills the performance.

Does your model have depthwise convolutions?